refactor: 타입 관련 코드 위치 수정
