+++ src/features/menu/hooks/useMenuSelectHook.tsx
@@ -1,25 +1,5 @@
 import { create } from 'zustand'
-
-interface MenuSelectState {
-  menuPrice: number | null
-  setMenuPrice: (menuPrice: number) => void
-  menuCount: number
-  setMenuCount: (n: number, menuPrice: number) => void
-  optionGroups: OptionGroup[]
-  setOptionGroups: (optionGroups: OptionGroup[]) => void
-}
-
-interface OptionGroup {
-  optionGroupId: number
-  optionGroupName: string
-  options: Option[]
-}
-
-interface Option {
-  optionId: number
-  optionName: string
-  optionPrice: number
-}
+import { MenuSelectState, OptionGroup } from '../types'
 
 export const useMenuSelectStore = create<MenuSelectState>((set) => ({
   menuPrice: null,
