+++ src/features/menu/types/index.ts
@@ -10,3 +10,24 @@ export interface OptionState {
   optionPrice: number
   setOptionPrice: (n: number) => void
 }
+
+export interface MenuSelectState {
+  menuPrice: number | null
+  setMenuPrice: (menuPrice: number) => void
+  menuCount: number
+  setMenuCount: (n: number, menuPrice: number) => void
+  optionGroups: OptionGroup[]
+  setOptionGroups: (optionGroups: OptionGroup[]) => void
+}
+
+export interface OptionGroup {
+  optionGroupId: number
+  optionGroupName: string
+  options: Option[]
+}
+
+export interface Option {
+  optionId: number
+  optionName: string
+  optionPrice: number
+}
