+++ src/features/reviews/api/index.ts
@@ -1,16 +1,8 @@
 import axiosInst from '@/util/axiosInst'
-import { Review, ReviewRequest, ReviewResponse } from '../types'
+import { ReviewListResponse, ReviewRequest, ReviewResponse } from '../types'
 
 export const getReviews = async (storeId: string, pageParam: number) => {
-  const res = await axiosInst.get<{
-    response: {
-      storeName: string
-      reviews: Review[]
-      page: number
-      size: number
-      totalCount: number
-    }
-  }>(`/store/${storeId}/reviews?page=${pageParam}`)
+  const res = await axiosInst.get<ReviewListResponse>(`/store/${storeId}/reviews?page=${pageParam}`)
 
   const { page, size, totalCount } = res.data.response
 
@@ -23,7 +15,7 @@ export const getReviews = async (storeId: string, pageParam: number) => {
 }
 
 export const postReview = async (storeId: string, review: ReviewRequest) => {
-  const res = await axiosInst.post<{ response: ReviewResponse }>(`/order/${storeId}/review`, review)
+  const res = await axiosInst.post<ReviewResponse>(`/order/${storeId}/review`, review)
 
   return res.data.response
 }
